requirements.txt
asgiref==3.5.0
Django==4.0.2
django-phone-field==1.8.1
django-phonenumber-field==6.1.0
phonenumbers==8.12.43
Pillow==9.0.1
sqlparse==0.4.2
tzdata==2021.5
